% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_duration.R
\name{add_duration}
\alias{add_duration}
\title{Creates a duration variable using the start and end time of the survey}
\usage{
add_duration(
  .dataset,
  duration_var_name = "duration",
  uuid = "_uuid",
  start = "start",
  end = "end"
)
}
\arguments{
\item{.dataset}{the clean dataframe}

\item{duration_var_name}{the name of the duration variable that will be created.}

\item{uuid}{the unique identifier of the survey}

\item{start}{start time of the survey, in KOBO format (e.g "2021-07-14T14:02:24.955+03:00")}

\item{end}{end time of the survey, in KOBO format (e.g "2021-07-14T14:03:28.955+03:00")}
}
\value{
the clean dataset with 6 new columns: duration, start_date, start_time, end_date, end_time, days_diff
}
\description{
Creates a duration variable using the start and end time of the survey
}
