% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_composite.R
\name{add_hhs}
\alias{add_hhs}
\title{Add the household hunger scale to the dataset}
\usage{
add_hhs(
  dataset,
  new_colname = "hhs",
  new_category_label = c("none", "little", "moderate", "severe", "very_severe"),
  hhs1a = "fs_hhs_nofood_yn",
  hhs1b = "fs_hhs_nofood_freq",
  hhs2a = "fs_hhs_sleephungry_yn",
  hhs2b = "fs_hhs_sleephungry_freq",
  hhs3a = "fs_hhs_daynoteating_yn",
  hhs3b = "fs_hhs_daynoteating_freq",
  yesno_choice = c("yes", "no"),
  frequency_choice = c("rarely_1_2", "sometimes_3_10", "often_10_times")
)
}
\arguments{
\item{dataset}{A dataframe}

\item{new_colname}{The prefix for the new columns. It has to be a string.}

\item{new_category_label}{A vector of 5 strings that will replace the default labels
c("none", "little", "moderate", "severe", "very_severe")}

\item{hhs1a}{The name of the column "In the past 4 weeks (30 days), was there ever no food to eat
of any kind in your house because of lack of resources to get food?". It has to be a string.}

\item{hhs1b}{The name of the column "How often did this happen in the past (4 weeks/30 days)?".
It has to be a string.}

\item{hhs2a}{The name of the column "In the past 4 weeks (30 days), did you or any household
member go to sleep at night hungry because there was not enough food?". It has to be a string.}

\item{hhs2b}{The name of the column "How often did this happen in the past (4 weeks/30 days)?".
It has to be a string.}

\item{hhs3a}{The name of the column "In the past 4 weeks (30 days), did you or any household
member go a whole day and night without eating anything at all because there was not enough food?".
It has to be a string.}

\item{hhs3b}{The name of the column "How often did this happen in the past (4 weeks/30 days)?".
It has to be a string.}

\item{yesno_choice}{A vector of 2 strings from the choices for the yes/no set of questions in
this order: c("yes", "no")}

\item{frequency_choice}{A vector of 3 strings from the choices for the frequency set of questions
in this order:  c("rarely_1_2", "sometimes_3_10", "often_10_times")}
}
\value{
It returns the dataframe with 5 extras columns: score for the 3 sets of questions
(from 0 to 2), the HHS score (from 0 to 6) and the HHS category.
}
\description{
Adds the 6 columns used for the Household Hunger Scale (HHS) to a dataframe.
}
