% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_percentage_missing.R
\name{add_percentage_missing}
\alias{add_percentage_missing}
\title{Adds the percentage of missing values per row}
\usage{
add_percentage_missing(
  .dataset,
  col_name = "percentage_missing",
  kobo_survey = NULL,
  type_to_include = c("integer", "date", "text", "select_one", "select_multiple")
)
}
\arguments{
\item{.dataset}{A dataset to add the percentage of missing values}

\item{col_name}{string variable with the name of the new column to be created, default is percentage_missing}

\item{kobo_survey}{KOBO question or survey sheet including at least the type and name.}

\item{type_to_include}{Types (from KOBO) to be included in the columns default are
integer, date, text, select_one and select_multiple.}
}
\value{
a dataset with the one additional column with the percentage of missing value
}
\description{
Adds the percentage of missing values per row
}
\examples{
data_test <- data.frame(
  uuid = c(1:3),
  col_1 = c(1:3),
  col_2 = c(NA, NA, "expenditures"),
  col_3 = c("with need", NA, "with need"),
  col_4 = c("food health school", NA, "food"),
  col_4.food = c(1, NA, 1),
  col_4.health = c(1, NA, 0),
  col_4.school = c(1, NA, 0)
)
kobo_survey <- data.frame(
  type = c("uuid",
           "integer",
           "select_one choice2",
           "select_one choice3",
           "select_multiple choice4"),
  name = c("uuid", "col_1", "col_2", "col_3", "col_4")
)
data_test \%>\% add_percentage_missing(kobo_survey = kobo_survey,
type_to_include = c("integer","select_one","select_multiple"))
data_test \%>\% add_percentage_missing()
}
\seealso{
\code{\link[=check_percentage_missing]{check_percentage_missing()}}
}
