% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_set_documentation.R
\docType{data}
\name{analysis_by_group}
\alias{analysis_by_group}
\title{Analysis by population group}
\usage{
analysis_by_group
}
\description{
Survey data
}
\details{
These data sets include HH data (raw and clean data) and analysis along with cleaning log.
}
\examples{
analysis_by_group
raw_data
cleaning_log
clean_data
survey
choices
overall_analysis
}
\keyword{datasets}
