% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recreate_parent_column.R
\name{auto_detect_sm_parents}
\alias{auto_detect_sm_parents}
\title{Detect select multiple parent columns}
\usage{
auto_detect_sm_parents(dataset, sm_separator = ".")
}
\arguments{
\item{dataset}{dataset to correct}

\item{sm_separator}{Separator for choice multiple questions. The default is "."
If using read_csv to read in data the separator will most likely be '/' where as if using read.csv it will likely be '.'}
}
\value{
a list of select multiple parent columns in data set.
}
\description{
\code{auto_detect_sm_parents} is mean to detect select multiple parent columns in a way that does
not rely on the XLSForm as the input
}
