% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recreate_parent_column.R
\name{auto_sm_parent_child}
\alias{auto_sm_parent_child}
\alias{auto_sm_parent_children}
\title{detect and group together select multiple parent and children columns}
\usage{
auto_sm_parent_children(dataset, sm_separator = ".")
}
\arguments{
\item{dataset}{dataset to correct}

\item{sm_separator}{Separator for choice multiple questions. The default is "."
If using read_csv to read in data the separator will most likely be '/' where as if using read.csv it will likely be '.'}
}
\value{
a data frame containing the the child select multiple columns alongside there parents and
a log with all changes recorded.
}
\description{
\code{auto_sm_parent_children} is mean to detect select multiple parent columns & children columns in a way that does
not rely on the XLSForm as the input
}
