% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/review_cleaning_log.R
\name{check_deletion_log}
\alias{check_deletion_log}
\title{Check deletion log}
\usage{
check_deletion_log(
  raw_data,
  clean_data,
  deletion_log,
  raw_data_uuid,
  clean_data_uuid,
  deletion_log_uuid
)
}
\arguments{
\item{raw_data}{Raw dataset}

\item{clean_data}{Clean dataset}

\item{deletion_log}{deletion log}

\item{raw_data_uuid}{Unique ID column name of raw dataset}

\item{clean_data_uuid}{Unique ID column name of clean dataset}

\item{deletion_log_uuid}{Unique ID column name of deletion log}
}
\value{
Discrepancy in deletion log and data sets
}
\description{
Check deletion log
}
