% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_duplicate.R
\name{check_duplicate}
\alias{check_duplicate}
\title{Checks for duplicated uuid}
\usage{
check_duplicate(.dataset, .col_to_check = "_uuid")
}
\arguments{
\item{.dataset}{a dataset to be check as a dataframe or a list with the
dataframe stored as "checked_dataset"}

\item{.col_to_check}{string character with the name of the uuid column}
}
\value{
return a list with the dataset checked stored as checked_dataset and
a dataframe with the duplicate log
}
\description{
Checks for duplicated uuid
}
\examples{
  testdata <- data.frame(
uuid = c(letters[1:4], "a", "b", "c"),
col_a = runif(7),
col_b = runif(7)
) \%>\%
  dplyr::rename(`_uuid` = uuid)

check_duplicate(testdata)

}
