% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_fcs.R
\name{check_fcs}
\alias{check_fcs}
\title{FCS component checks}
\usage{
check_fcs(
  dataset,
  uuid = "X_uuid",
  cereals = "fs_fcs_cerealgrainroottuber",
  pulses = "fs_fcs_beansnuts",
  dairy = "fs_fcs_dairy",
  meat = "fs_fcs_meatfishegg",
  vegetables = "fs_fcs_vegetableleave",
  fruits = "fs_fcs_fruit",
  oil = "fs_fcs_fat",
  sugar = "fs_fcs_sugar"
)
}
\arguments{
\item{dataset}{Data set}

\item{uuid}{Column containing unique identifier}

\item{cereals}{Variable name for- In the last 7 days, on how many days did your household eat cereals, grains, roots and tubers, including wild roots?}

\item{pulses}{Variable name for-  In the last 7 days, on how many days did your household eat any beans / legumes, pulses or nuts?}

\item{dairy}{Variable name for- In the last 7 days, on how many days did your household drink milk or eat other dairy products?}

\item{meat}{Variable name for-  In the last 7 days, on how many days did your household eat meat, fish or eggs?}

\item{vegetables}{Variable name for- In the last 7 days, on how many days did your household eat vegetables or leaves, including all wild vegetables and leaves?}

\item{fruits}{Variable name for- In the last 7 days, on how many days did your household eat fruits, including all wild fruits?}

\item{oil}{Variable name for- In the last 7 days, on how many days did your household eat oil, fat, or butter?}

\item{sugar}{Variable name for- In the last 7 days, on how many days did your household eat sugar or sugary foods?}
}
\value{
dataset with potential issues
}
\description{
FCS component checks
}
