% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_for_pii.R
\name{check_for_pii}
\alias{check_for_pii}
\title{Checks for potential PII}
\usage{
check_for_pii(df, element_name = NULL, uuid = "X_uuid", words_to_look = NULL)
}
\arguments{
\item{df}{Data set}

\item{element_name}{If the input is a list file, please specify the element name that contains the dataset}

\item{uuid}{Column name defining the uuid}

\item{words_to_look}{Specify the words that might be the PIIs}
}
\value{
dataset with potential PII
}
\description{
Checks for potential PII
}
