% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_for_value.R
\name{check_for_value}
\alias{check_for_value}
\title{Check for value(s) in the dataset}
\usage{
check_for_value(
  df,
  uuid_col_name,
  element_name = "checked_dataset",
  values_to_look = c(99, 999, 999, 88, 888, 888)
)
}
\arguments{
\item{df}{data frame or a list}

\item{uuid_col_name}{UUID}

\item{element_name}{Name of the dataset in list}

\item{values_to_look}{Values to look. Default are- 99,999,999,88,888,888}
}
\value{
A dataframe as cleaning log format
}
\description{
Check for value(s) in the dataset
}
\examples{
df <- data.frame(
X_uuid = paste0("uuid_",1:100),
age = c(sample(18:80,replace = TRUE,size = 96),99,99,98,88),
gender = c("99",sample(c("male","female"),replace = TRUE,size = 95),"98","98","88","888"))
check_for_value(df = df,uuid_col_name = "X_uuid",element_name = "checked_dataset",values_to_look = c(99,98,88,888))

}
