% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_others.R
\name{check_others}
\alias{check_others}
\title{Generate a log for other follow up questions}
\usage{
check_others(dataset, uuid = "uuid", var_list = NULL)
}
\arguments{
\item{dataset}{dataset}

\item{uuid}{uuid column}

\item{var_list}{list of questions to generate the logs for}
}
\value{
a log with all the values that may require recoding
}
\description{
Generate a log for other follow up questions
}
