% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_others.R
\name{check_others_checks}
\alias{check_others_checks}
\title{Check if the input passed to the check_others function is correct}
\usage{
check_others_checks(dataset, uuid, var_list)
}
\arguments{
\item{dataset}{dataset}

\item{uuid}{uuid column}

\item{var_list}{list of questions to generate the logs for}
}
\value{
nothing
}
\description{
Check if the input passed to the check_others function is correct
}
