% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_outliers.R
\name{check_outliers}
\alias{check_outliers}
\title{check outliers over the dataset}
\usage{
check_outliers(
  df,
  kobo_tool_location = NULL,
  uuid_col_name,
  element_name = "checked_dataset",
  columns_to_add = NULL,
  strongness_factor = 3,
  minimum_unique_value_of_variable = NULL,
  columns_to_remove = NULL
)
}
\arguments{
\item{df}{data frame or a list}

\item{kobo_tool_location}{kobo tool path. Default is null}

\item{uuid_col_name}{uuid}

\item{element_name}{name of the dataset in list}

\item{columns_to_add}{Variables those must be included in the output}

\item{strongness_factor}{Strongness factor define how strong your outliers will be. The default is 1.5.}

\item{minimum_unique_value_of_variable}{Default is NULL, mean this parameter won't be considered. For example 10 means for any variable where number of unique value is less than 10, then the variable won't be considered for outlier checking.}

\item{columns_to_remove}{Variables those must not consider for outlier checks even though they are numeric}
}
\value{
Outliers
}
\description{
check outliers over the dataset
}
\examples{
df_outlier<- data.frame(
 uuid = paste0("uuid_", 1:100),
 one_value = c(round(runif(90, min = 45,max =55)), round(runif(5)), round(runif(5,99,100))),
 expense = c(sample(200:500,replace = TRUE,size = 95),c(600,100,80,1020,1050)),
 income = c(c(60,0,80,1020,1050),sample(20000:50000,replace = TRUE,size = 95)),
 yy = c(rep(100,99),10)
)

check_outliers(df = df_outlier,uuid_col_name = "uuid")
}
