% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_pii.R
\name{check_pii}
\alias{check_pii}
\title{Checks for potential PII}
\usage{
check_pii(
  dataset,
  element_name = "checked_dataset",
  uuid_column = "uuid",
  words_to_look = NULL
)
}
\arguments{
\item{dataset}{dataset to be check as a dataframe or a list with the dataframe stored as "checked_dataset"}

\item{element_name}{If the input is a list file, please specify the element name that contains the dataset}

\item{uuid_column}{uuid column in the dataset. Default is uuid.}

\item{words_to_look}{Specify the words that might be the PIIs}
}
\value{
dataset with potential PII
}
\description{
Checks for potential PII
}
\examples{
check_pii(cleaningtools_raw_data, uuid_column = "X_uuid")

}
