% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_soft_duplicates.R
\name{check_soft_duplicates}
\alias{check_soft_duplicates}
\title{Checks for survey similarities - Soft Duplicates}
\usage{
check_soft_duplicates(
  dataset,
  kobo_survey,
  uuid = "uuid",
  idnk_value = "idnk",
  sm_seperator = ".",
  log_name = "soft_duplicate_log"
)
}
\arguments{
\item{dataset}{RawData to be checked as a dataframe or a list with dataframe
stored as "checked_dataset"}

\item{kobo_survey}{Kobo Survey Sheet}

\item{uuid}{String character for the uuid column. By default "uuid"}

\item{idnk_value}{String character for the value of the "I don't know" value}

\item{sm_seperator}{Separator for choice multiple questions. The default is "."}

\item{log_name}{Name of the log dataframe flagged in the end of the function}
}
\value{
return a list with the dataset checked stored as checked_dataset and
a dataframe with the soft duplicate log
}
\description{
Checks for survey similarities - Soft Duplicates
}
\examples{
soft_duplicates <- check_soft_duplicates(dataset = cleaningtools_raw_data,
                                         kobo_survey = cleaningtools_survey,
                                         uuid = "X_uuid",
                                         idnk_value = "dont_know",
                                         sm_seperator = ".",
                                         log_name = "soft_duplicate_log")
}
