% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_value.R
\name{check_value}
\alias{check_value}
\title{Check for value(s) in the dataset}
\usage{
check_value(
  dataset,
  uuid_column = "uuid",
  element_name = "checked_dataset",
  values_to_look = c(99, 999, 999, 88, 888, 888)
)
}
\arguments{
\item{dataset}{dataset to be check as a dataframe or a list with the dataframe stored as "checked_dataset"}

\item{uuid_column}{uuid column in the dataset. Default is "uuid".}

\item{element_name}{Name of the dataset in list}

\item{values_to_look}{Values to look. Default are- 99,999,999,88,888,888}
}
\value{
A dataframe as cleaning log format
}
\description{
Check for value(s) in the dataset
}
\examples{
df <- data.frame(
  X_uuid = paste0("uuid_", 1:100),
  age = c(sample(18:80, replace = TRUE, size = 96), 99, 99, 98, 88),
  gender = c("99", sample(c("male", "female"),
    replace = TRUE, size = 95
  ), "98", "98", "88", "888")
)
check_value(
  dataset = df, uuid_column = "X_uuid",
  element_name = "checked_dataset",
  values_to_look = c(99, 98, 88, 888)
)

}
