% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/review_cleaning_log.R
\name{cleaning_log_check}
\alias{cleaning_log_check}
\title{Compare cleaning logs}
\usage{
cleaning_log_check(
  raw_data = raw_data_hh_2,
  raw_data_uuid = "X_id",
  clean_data = clean_data_hh,
  clean_data_uuid = "X_id",
  cleaning_log = cleaning_log_only,
  cleaning_log_uuid = "id",
  cleaning_log_old_value = "old_value",
  cleaning_log_new_value = "new_value",
  cleaning_log_question_name = "question.name",
  deletion_log = deletion_log_hh,
  deletion_log_uuid = "id",
  stop_comparing_raw_value = T
)
}
\arguments{
\item{raw_data}{Raw dataset}

\item{raw_data_uuid}{Unique ID column name of raw dataset}

\item{clean_data}{Clean dataset}

\item{clean_data_uuid}{Unique ID column name of clean dataset}

\item{cleaning_log}{Cleaning log}

\item{cleaning_log_uuid}{Unique ID column name of cleaning log}

\item{cleaning_log_old_value}{Cleaning log old value}

\item{cleaning_log_new_value}{Cleaning log new value}

\item{cleaning_log_question_name}{Column name of cleaning log storing the dataset column name}

\item{deletion_log}{deletion log}

\item{deletion_log_uuid}{Unique ID column name of deletion log}

\item{stop_comparing_raw_value}{TRUE to stop the checks for discrepancy between old_value and raw value}
}
\value{
Discrepancy in cleaning log
}
\description{
Compare cleaning logs
}
