% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaningtools-package.R
\docType{package}
\name{cleaningtools-package}
\alias{cleaningtools}
\alias{cleaningtools-package}
\title{cleaningtools: cleaningtools package focuses on data cleaning}
\description{
The cleaningtools package focuses on cleaning, and has three components: **Check**, which includes a set of functions that flag values, such as check_outliers and check_logical. **Create**, which includes a set of functions to create different items for use in cleaning, such as the cleaning log from the checks, clean data, and enumerator performance. **Review**, which includes a set of functions to review the cleaning.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://impact-initiatives.github.io/cleaningtools/}
}

}
\author{
\strong{Maintainer}: Yann Say \email{yann.say@impact-initiatives.org} (\href{https://orcid.org/0000-0002-7390-4209}{ORCID})

Authors:
\itemize{
  \item Mehedi Khan \email{mh.khan@reach-initiative.org} (\href{https://orcid.org/0000-0002-4276-4485}{ORCID})
}

}
\keyword{internal}
