% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_set_documentation.R
\docType{data}
\name{cleaningtools_analysis_by_group}
\alias{cleaningtools_analysis_by_group}
\title{Analysis by population group}
\usage{
cleaningtools_analysis_by_group
}
\description{
Survey data
}
\details{
These data sets include HH data (raw and clean data) and analysis along with cleaning log.
}
\examples{
cleaningtools_analysis_by_group
cleaningtools_raw_data
cleaningtools_cleaning_log
cleaningtools_clean_data
cleaningtools_survey
cleaningtools_choices
cleaningtools_overall_analysis
cleaningtools_sample_frame
}
\keyword{datasets}
