% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_set_documentation.R
\docType{data}
\name{cleaningtools_choices}
\alias{cleaningtools_choices}
\title{Choices tab of kobo tool}
\usage{
cleaningtools_choices
}
\description{
Choices tab of kobo tool
}
\keyword{datasets}
