% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_set_documentation.R
\docType{data}
\name{cleaningtools_food_consumption_df}
\alias{cleaningtools_food_consumption_df}
\title{Dataset with food consumption, household hunger Score component}
\usage{
cleaningtools_food_consumption_df
}
\description{
Dataset with food consumption, household hunger Score component
}
\keyword{datasets}
