% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{coerce_to_character}
\alias{coerce_to_character}
\title{Coerce numeric values to character, without scientific noting and NA are kept as NA.}
\usage{
coerce_to_character(x)
}
\arguments{
\item{x}{a value to convert}
}
\value{
x as character
}
\description{
Coerce numeric values to character, without scientific noting and NA are kept as NA.
}
\examples{

coerce_to_character(c("a", NA))
coerce_to_character(c(1, 10000000, NA))

}
