% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_cleaning_logs.R
\name{combine_cleaning_log}
\alias{combine_cleaning_log}
\title{Merging the cleaning logs}
\usage{
combine_cleaning_log(list_of_log = a, dataset_name = "checked_dataset")
}
\arguments{
\item{list_of_log}{A list file containig all the logs}

\item{dataset_name}{The dataset's name, if it exists in the list_of_log}
}
\value{
Combined cleaning log
}
\description{
Merging the cleaning logs
}
\examples{
  list <- cleaningtools::cleaningtools_raw_data |> check_for_pii() |>
          check_duplicate(uuid_col_name = "X_uuid") |>
          check_for_value(uuid_col_name = "X_uuid")

 combine_cleaning_log(list_of_log = list)


}
