% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_audits.R
\name{create_audit_list}
\alias{create_audit_list}
\title{Read all audit files from a zip}
\usage{
create_audit_list(audit_zip_path, dataset = NULL, uuid_column = "uuid")
}
\arguments{
\item{audit_zip_path}{location to .zip file. It must contain .zip}

\item{dataset}{dataset to (optional)}

\item{uuid_column}{uuid column if a dataset is provided. It will only read the uuid
present in the dataset. It will only be used if a dataset is provided.}
}
\value{
A list with all the audits file read as they are.
}
\description{
Read all audit files from a zip
}
\examples{
\dontrun{
create_audit_list("audit_path.zip")
}
}
