% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_clean_data.R
\name{create_clean_data}
\alias{create_clean_data}
\title{implement cleaning log on raw data set.}
\usage{
create_clean_data(
  df,
  df_uuid,
  cl,
  cl_change_type_col,
  values_for_change_response = "change_response",
  values_for_blank_response = "blank_response",
  values_for_no_change = "no_action",
  values_for_remove_survey = "remove_survey",
  cl_change_col,
  cl_uuid,
  cl_new_val
)
}
\arguments{
\item{df}{raw data (data frame or list)}

\item{df_uuid}{column in raw data with uuid}

\item{cl}{cleaning + deletion log (data.frame).}

\item{cl_change_type_col}{column in cleaning log which specifies change type to be made}

\item{values_for_change_response}{values in change type column which should be changed to a new value.}

\item{values_for_blank_response}{values in change type column which should be blank (NA).}

\item{values_for_no_change}{values in change type column which should NOT be changed to a new value.}

\item{values_for_remove_survey}{values in change type column which should be deleted from the data.}

\item{cl_change_col}{column in cleaning log which specifies data set column to change}

\item{cl_uuid}{uuid in cleaning log}

\item{cl_new_val}{cleaning log column specifying the new correct value}
}
\value{
clean data set
}
\description{
implement cleaning log on raw data set.
}
\examples{
cleaning_log_test <- data.frame(
uuid = paste0("uuid",1:4),
question= c("age","gender","pop_group","strata"),
change_type = c("blank_response","no_change","Delete","change_res"),
new_value = c(NA_character_,NA_character_,NA_character_,"st-a")
)
test_data <- data.frame(
uuid =  paste0("uuid",1:4),
age = c(180,23,45,67),
gender = c("male","female","male","female"),
pop_group = c("idp","refugee","host","idp"),
strata = c("a","b","c","d")
)


check_cleaning_log(df = test_data,df_uuid = "uuid",cl =cleaning_log_test,
                cl_change_type_col = "change_type",
                values_for_change_response =  "change_res",
                cl_change_col = "question",cl_uuid = "uuid",cl_new_val = "new_value")

create_clean_data(df = test_data,df_uuid = "uuid",cl = cleaning_log_test,
               cl_change_type_col =  "change_type",
               values_for_change_response = "change_res",
               values_for_blank_response = "blank_response",
               values_for_no_change = "no_change",
               values_for_remove_survey = "Delete",
               cl_change_col =  "question",
               cl_uuid = "uuid",
               cl_new_val = "new_value" )
}
