% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_and_compare_cl_log.R
\name{create_cleaning_log}
\alias{create_cleaning_log}
\title{Generates cleaning log}
\usage{
create_cleaning_log(
  raw_data = raw_data,
  raw_data_uuid = "X_id",
  clean_data = clean_data,
  clean_data_uuid = "X_id",
  check_for_deletion_log = T,
  check_for_variable_name = T
)
}
\arguments{
\item{raw_data}{Raw dataset}

\item{raw_data_uuid}{Unique ID column name of raw dataset}

\item{clean_data}{Clean dataset}

\item{clean_data_uuid}{Unique ID column name of clean dataset}

\item{check_for_deletion_log}{TRUE to flag the removed survey}

\item{check_for_variable_name}{TRUE to flag the removed variables}
}
\value{
Cleaning log
}
\description{
Generates cleaning log
}
