% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cleaning_logbook.R
\name{create_cleaning_logbook}
\alias{create_cleaning_logbook}
\title{Create data_exstract, logbook, deletion_log tables and add them to the cleaning logbook template}
\usage{
create_cleaning_logbook(
  raw_data,
  cleaning_log,
  uuid = "X_uuid",
  col_enum = NA,
  filename = NA
)
}
\arguments{
\item{raw_data}{Raw Dataset}

\item{cleaning_log}{Cleaning Log}

\item{uuid}{}

\item{col_enum}{Enumerator column name in both raw_data and cleaning_log. Must be the same and by default NA}

\item{filename}{URL link for the location of the updated cleaning logbook}
}
\value{
Return a list of 3 tables:
- "template_data_extract"
- "template_logbook"
- "template_deletion_log"
And update the cleaning logbook with the new tables and automatically update the enumerator performance table
}
\description{
Create data_exstract, logbook, deletion_log tables and add them to the cleaning logbook template
}
\examples{
\dontrun{
create_enumerator_performance(
raw_data = cleaningtools::cleaningtools_raw_data,
cleaning_log = cleaningtools::cleaningtools_cleaning_log,
col_enum = "enumerator_num",
filename = "new_cleaning_logbook_draft.xlsx"
)}
}
