% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_col_range}
\alias{create_col_range}
\title{Generate excel range to be used for the data validation formula in excel}
\usage{
create_col_range(variable, data.val)
}
\arguments{
\item{variable}{The name of the column/question for which the Excel range is to be generated.}

\item{data.val}{A dataframe containing the column specified by \code{variable}.}
}
\value{
A string representing the Excel cell range containing all choices for the column \code{variable}
in the sheet 'validation_rules'.
}
\description{
This function returns an Excel cell range for a given column in a dataframe.
The range is formatted for use in Excel functions or formulas and is prefixed
with the sheet name 'validation_rules'.
}
\examples{
# Assume df is a sample dataframe with a column named 'consent'.
# range_string <- create_col_range('consent', df)

}
