% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_audits.R
\name{create_duration_from_audit_sum_all}
\alias{create_duration_from_audit_sum_all}
\title{Calculate duration from audit summing all time}
\usage{
create_duration_from_audit_sum_all(audit_file)
}
\arguments{
\item{audit_file}{a dataframe with a single audit file,
it needs start, end, node column}
}
\value{
A dataframe with the duration in ms and duration in minutes.
}
\description{
Duration is the difference between the sum of all start and the sum of all end.
It ignores all node that are empty such as form_start, end screen, form save, etc.
}
\examples{
some_audit <- data.frame(
  event = c("form start", rep("question", 5)),
  node = c("", paste0("/xx/question", 1:5)),
  start = c(
    1661415887295, 1661415887301, 1661415890819,
    1661415892297, 1661415893529, 1661415894720
  ),
  end = c(
    NA, 1661415890790, 1661415892273,
    1661415893506, 1661415894703, 1661415896452
  )
)
create_duration_from_audit_sum_all(some_audit)
}
