% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_enumerator_performance.R
\name{create_enumerator_performance}
\alias{create_enumerator_performance}
\title{Create enumerator performance tables and add them to the cleaning logbook}
\usage{
create_enumerator_performance(
  raw_data,
  cleaning_log,
  col_enum = NA,
  issue_col = "issue",
  link_cleaning_logbook = NA,
  filename = NA
)
}
\arguments{
\item{raw_data}{Raw Dataset}

\item{cleaning_log}{Cleaning Log}

\item{col_enum}{Enumerator column name in both raw_data and cleaning_log. Must be the same and by default NA}

\item{issue_col}{Name of the issue column in the cleaning log. By Default "issue"}

\item{link_cleaning_logbook}{URL link to the RDD cleaning logbook template containing the sheet "04_Enumerator - performance"}

\item{filename}{URL link for the location of the updated cleaning logbook}
}
\value{
Return a list of 5 named tables:
- "Number of surveys collected by enumerators"
- "Number of changes by enumerators"
- "Number of changes by enumerators filtered by issues"
- "Number of deletions by enumerators"
- "Number of deletions by enumerators"
And update the cleaning logbook with the new tables
}
\description{
Create enumerator performance tables and add them to the cleaning logbook
}
\examples{
\dontrun{
create_enumerator_performance(
raw_data = cleaningtools::cleaningtools_raw_data,
cleaning_log = cleaningtools::cleaningtools_cleaning_log,
col_enum = "enumerator_num", issue_col = "issue",
link_cleaning_logbook = "Research_cycle_data_cleaning_logbook_template_v3_202208_FINAL.xlsx",
filename = "new_cleaning_logbook_draft.xlsx"
)}
}
