% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_xlsx_cleaning_log.R
\name{create_formated_wb}
\alias{create_formated_wb}
\title{Creates formatted workbook with openxlsx}
\usage{
create_formated_wb(
  write_list,
  column_for_color = NULL,
  header_front_size = 12,
  header_front_color = "#FFFFFF",
  header_fill_color = "#ee5859",
  header_front = "Arial Narrow",
  body_front = "Arial Narrow",
  body_front_size = 11
)
}
\arguments{
\item{write_list}{List of dataframe}

\item{column_for_color}{Column name in the dataframe which should be used for colorizing the cell. The default is null.}

\item{header_front_size}{Header front size (default is 12)}

\item{header_front_color}{Hexcode for header front color (default is white)}

\item{header_fill_color}{Hexcode for header fill color (default is red)}

\item{header_front}{Front name for header (default is Arial Narrow)}

\item{body_front}{Front name for body (default is Arial Narrow)}

\item{body_front_size}{Front size for body (default is 11)}
}
\value{
A workbook
}
\description{
Creates formatted workbook with openxlsx
}
\examples{
checks_list <- cleaningtools::cleaningtools_raw_data |>
  check_pii(uuid_column = "X_uuid") |>
  check_duplicate(uuid_column = "X_uuid") |>
  check_value(uuid_column = "X_uuid")

create_combined_log(list_of_log = checks_list) |>
  create_formated_wb()
}
