% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_formatted_choices}
\alias{create_formatted_choices}
\title{Format and filter Choices for 'select_one' Questions}
\usage{
create_formatted_choices(choices, tool)
}
\arguments{
\item{choices}{A dataframe containing the choices sheet from the Kobo tool
Expected to have at least the columns \code{list_name} and \code{name}.}

\item{tool}{A dataframe containing the survey sheet from the Kobo tool
Expected to have at least the columns \code{type} and \code{name}.}
}
\value{
A dataframe containing 'select_one' questions with their corresponding choices.
The dataframe has columns from both the \code{tool} and \code{choices} dataframes.
}
\description{
This function returns a dataframe containing 'select_one' questions
from the \code{tool} dataframe with their corresponding choices in the \code{choices} dataframe.
}
\examples{
# formatted_choices <- create_formatted_choices(choices, tool)

}
