% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_logic_for_other.R
\name{create_logic_for_other}
\alias{create_logic_for_other}
\title{Create logical checks for "other" values.}
\usage{
create_logic_for_other(
  kobo_survey,
  sm_separator = ".",
  compare_with_dataset = FALSE,
  dataset = NULL
)
}
\arguments{
\item{kobo_survey}{Kobo survey sheet.}

\item{sm_separator}{Separator for choice multiple questions. The default is "."}

\item{compare_with_dataset}{Default FALSE will not compare the applicability of logics with the
dataset whereas TRUE will check the applicability of logics.}

\item{dataset}{dataset to be checked.}
}
\value{
A dataframe will all logical tests to check "others" value. It should be used in
combination of review_others
}
\description{
Create logical checks for "other" values.
}
\examples{
create_logic_for_other(
  kobo_survey = cleaningtools::cleaningtools_survey,
  sm_separator = ".",
  dataset = cleaningtools::cleaningtools_clean_data,
  compare_with_dataset = TRUE
)

}
