% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_uuid.R
\name{detect_uuid}
\alias{detect_uuid}
\title{Helper to look to guess uuid name}
\usage{
detect_uuid(dataset, name_to_search = "uuid")
}
\arguments{
\item{dataset}{dataset to look from}

\item{name_to_search}{pattern to search, default is "uuid"}
}
\value{
text vector with the name to search.
If there is 0 or 2 or more possible ansers, it will
give an error.
}
\description{
Helper to look to guess uuid name
}
\examples{
test1_df <- data.frame(uuid = letters,
                          x = runif(26),
                          hello = LETTERS)
detect_uuid(test1_df)
}
