% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{detect_variable}
\alias{detect_variable}
\title{detects variables names in code}
\usage{
detect_variable(codeintext)
}
\arguments{
\item{codeintext}{a string vector to look for variables}
}
\value{
a vector with strings
}
\description{
detects variables names in code
}
\examples{
string_to_check <- c(
  "ki_age> 1",
  "noo==\"mar\"",
  "sum(var1, var2, na.rm = T)",
  "oks != \"mar2\"",
  "oknospace!=\"mar2\""
)
detect_variable(string_to_check)
}
