% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_set_documentation.R
\docType{data}
\name{food_consumption_df}
\alias{food_consumption_df}
\title{Dataset with food consumption, household hunger Score component}
\usage{
food_consumption_df
}
\description{
Dataset with food consumption, household hunger Score component
}
\keyword{datasets}
