% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_big_table.R
\name{knit_big_table}
\alias{knit_big_table}
\title{Format table for quarto html output}
\usage{
knit_big_table(table_to_print, .height = "500px")
}
\arguments{
\item{table_to_print}{A data frame to print}

\item{.height}{Height in the html output}
}
\value{
Table with some features for quarto html output
}
\description{
Format table for quarto html output
}
\examples{
#knit_big_table(table_to_print)
}
\keyword{internal}
