% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_composite.R
\name{make_fcs}
\alias{make_fcs}
\title{Making FCS variables}
\usage{
make_fcs(
  dataset,
  var_name = "fcs",
  cereals = "fs_fcs_cerealgrainroottuber",
  pulses = "fs_fcs_beansnuts",
  dairy = "fs_fcs_dairy",
  meat = "fs_fcs_meatfishegg",
  vegetables = "fs_fcs_vegetableleave",
  fruits = "fs_fcs_fruit",
  oil = "fs_fcs_fat",
  sugar = "fs_fcs_sugar",
  threshold = c(28, 42)
)
}
\arguments{
\item{dataset}{Data set}

\item{var_name}{Intended FCS variable name. Note than for score variable name will be {given_name + _score} and for category, it will be {given_name + _cat}.}

\item{cereals}{Variable name for- In the last 7 days, on how many days did your household eat cereals, grains, roots and tubers, including wild roots?}

\item{pulses}{Variable name for-  In the last 7 days, on how many days did your household eat any beans / legumes, pulses or nuts?}

\item{dairy}{Variable name for- In the last 7 days, on how many days did your household drink milk or eat other dairy products?}

\item{meat}{Variable name for-  In the last 7 days, on how many days did your household eat meat, fish or eggs?}

\item{vegetables}{Variable name for- In the last 7 days, on how many days did your household eat vegetables or leaves, including all wild vegetables and leaves?}

\item{fruits}{Variable name for- In the last 7 days, on how many days did your household eat fruits, including all wild fruits?}

\item{oil}{Variable name for- In the last 7 days, on how many days did your household eat oil, fat, or butter?}

\item{sugar}{Variable name for- In the last 7 days, on how many days did your household eat sugar or sugary foods?}

\item{threshold}{Threshold for FC category. Default is c(28,42). Means anything less than 29 will be poor, 29 to 42 will be bordeline and greater than 42 will be acceptable.}
}
\value{
dataset with potential issues
}
\description{
Making FCS variables
}
