% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_set_documentation.R
\docType{data}
\name{overall_analysis}
\alias{overall_analysis}
\title{Nation/all population level analysis.}
\usage{
overall_analysis
}
\description{
Nation/all population level analysis.
}
\keyword{datasets}
