% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_log.R
\name{print_log}
\alias{print_log}
\title{Help to print log or message for the template}
\usage{
print_log(log_to_print, message)
}
\arguments{
\item{log_to_print}{A data frame to print}

\item{message}{A message to print if the data frame is empty}
}
\value{
The data frame to print or an alternative message if the dataframe is empty
}
\description{
Help to print log or message for the template
}
\examples{
#print_log(log_to_print, "No values flagged")
}
