% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_composite.R
\name{recoding_hhs}
\alias{recoding_hhs}
\title{Recode the Household Hunger Score set of questions}
\usage{
recoding_hhs(hhs_yesno, hhs_freq, yesno_choice, frequency_choice)
}
\arguments{
\item{hhs_yesno}{the yes/no column of a given set}

\item{hhs_freq}{the frequency column of the given set}

\item{yesno_choice}{A vector of 2 strings from the choices for the yes/no set of questions in
this order: c("yes", "no")}

\item{frequency_choice}{A vector of 3 strings from the choices for the frequency set of questions
in this order:  c("rarely_1_2", "sometimes_3_10", "often_10_times")}
}
\value{
A vector of the same length from 0 to 2.
}
\description{
This is a helper for the function add_hhs
}
