% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recreate_parent_column.R
\name{recreate_parent_column}
\alias{recreate_parent_column}
\title{This function recreates the concerted columns for select multiple questions}
\usage{
recreate_parent_column(
  df,
  uuid = "uuid",
  kobo_survey_sheet = NULL,
  kobo_choices_sheet = NULL,
  sm_sep = "."
)
}
\arguments{
\item{df}{data frame}

\item{uuid}{Unique ID column(Default is uuid).}

\item{kobo_survey_sheet}{Kobo survey tab}

\item{kobo_choices_sheet}{Kobo choices tab}

\item{sm_sep}{Separator between question and choice. Default is ..}
}
\description{
This function recreates the concerted columns for select multiple questions
}
\examples{
test_data <- dplyr::tibble(
uuid = paste0("uuid_",1:6),
gender = rep(c("male","female"),3),
reason = c("xx,yy","xx,zy",
          "zy","xx,xz,zy",
          NA_character_,"xz"),
reason.xx = c(0,1,0,1,0,0),
reason.yy = c(1,0,0,0,1,0),
reason.xz = c(0,0,0,1,0,1),
reason.zy = c(0,1,1,1,0,0),
reason_zy = c(NA_character_,"A","B","C",NA_character_,NA_character_))
recreate_parent_column(df = test_data, uuid = "uuid", sm_sep = ".")

}
