% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_clean_data.R
\name{review_cleaning_log}
\alias{review_cleaning_log}
\title{check cleaning log}
\usage{
review_cleaning_log(
  raw_dataset,
  raw_data_uuid_column = "uuid",
  cleaning_log,
  cleaning_log_uuid_column = "uuid",
  cleaning_log_question_column,
  cleaning_log_new_value_column,
  cleaning_log_change_type_column,
  change_response_value = "change_response"
)
}
\arguments{
\item{raw_dataset}{Raw dataset}

\item{raw_data_uuid_column}{uuid column in the raw dataset. Default is "uuid".}

\item{cleaning_log}{cleaning log (data.frame)}

\item{cleaning_log_uuid_column}{uuid column in the cleaning log. Default is "uuid".}

\item{cleaning_log_question_column}{column in cleaning log which specifies data set column to change}

\item{cleaning_log_new_value_column}{cleaning log column specifying the new correct value}

\item{cleaning_log_change_type_column}{column in cleaning log which specifies change type to be made}

\item{change_response_value}{values in change type column which should be changed to a new value.}
}
\value{
cleaning log with only problematic entries and note specifying problem
}
\description{
check cleaning log
}
