% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/review_others.R
\name{review_others}
\alias{review_others}
\title{Review discrepancy between kobo relevancies and the dataset.}
\usage{
review_others(
  dataset,
  uuid_column = "uuid",
  kobo_survey,
  sm_separator = ".",
  information_to_add = NULL,
  columns_not_to_check = NULL
)
}
\arguments{
\item{dataset}{dataset to be check}

\item{uuid_column}{uuid column in the dataset. Default is "uuid".}

\item{kobo_survey}{Kobo survey sheet.}

\item{sm_separator}{Separator for choice multiple questions. The default is "."}

\item{information_to_add}{string vector optional, if to add some information to the log (today,village etc.)}

\item{columns_not_to_check}{Columns to exclude from the checks}
}
\value{
Any discrepancy between kobo relevancies and the dataset.
}
\description{
Review discrepancy between kobo relevancies and the dataset.
}
\examples{
review_others(
  dataset = cleaningtools::cleaningtools_clean_data,
  uuid_column = "X_uuid", kobo_survey = cleaningtools_survey
)

}
