% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/review_sample_frame_with_dataset.R
\name{review_sample_frame_with_dataset}
\alias{review_sample_frame_with_dataset}
\title{Compares the sample frame with the clean data}
\usage{
review_sample_frame_with_dataset(
  sample_frame,
  sampling_frame_strata_column,
  sampling_frame_target_survey_column = "Total.no.of.HH",
  clean_dataset,
  clean_dataset_strata_column,
  consent_column = "consent",
  consent_yes_value = "yes"
)
}
\arguments{
\item{sample_frame}{A sample frame. It mush have a column for strata and another column defining the number of needed survey per strata.}

\item{sampling_frame_strata_column}{Column name for strata in the sample frame}

\item{sampling_frame_target_survey_column}{Column name defining number of survey per strata in sample frame.}

\item{clean_dataset}{Clean data}

\item{clean_dataset_strata_column}{Column name for strata in clean data}

\item{consent_column}{Dataset consent column}

\item{consent_yes_value}{Choice for consent yes}
}
\value{
a dataset with the two additional columns with the number of completed surveys and number of remaining survey.
}
\description{
Compares the sample frame with the clean data
}
\examples{
review_sample_frame_with_dataset(
  sample_frame = cleaningtools::cleaningtools_sample_frame,
  sampling_frame_strata_column = "Neighbourhood",
  sampling_frame_target_survey_column = "Total.no.of.HH",
  clean_dataset = cleaningtools::cleaningtools_clean_data,
  clean_dataset_strata_column = "neighbourhood",
  consent_column = "consent_remote",
  consent_yes_value = "yes"
)
}
