% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{verify_valid_survey}
\alias{verify_valid_survey}
\title{Verify if the Kobo survey dataframe is valid}
\usage{
verify_valid_survey(kobo_survey)
}
\arguments{
\item{kobo_survey}{A dataframe representing the Kobo survey sheet.}
}
\value{
Logical. Returns \code{TRUE} if the dataframe meets all the criteria. Otherwise, it returns \code{FALSE}.
}
\description{
This function checks whether the provided Kobo survey dataframe meets certain criteria:
\enumerate{
\item It is a dataframe.
\item It is not empty.
\item It contains columns named "type" and "name".
}
}
\examples{
# Assume df_valid is a valid Kobo survey dataframe
df_valid <- data.frame(type = c("integer", "selecte_one yesno"), name = c("age", "consent"))
verify_valid_survey(df_valid) # should return TRUE

# Assume df_invalid lacks the required columns
df_invalid <- data.frame(column1 = c("integer", "selecte_one yesno"), column2 = c("age", "consent"))
verify_valid_survey(df_invalid) # should return FALSE
}
