% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_audits.R
\name{check_duration}
\alias{check_duration}
\title{Check if duration is outside of a range}
\usage{
check_duration(
  dataset,
  column_to_check,
  uuid_column = "uuid",
  log_name = "duration_log",
  lower_bound = 25,
  higher_bound = 90
)
}
\arguments{
\item{dataset}{a dataset to be check as a dataframe or a list with the
dataframe stored as "checked_dataset"}

\item{column_to_check}{string character with the name of the duration column}

\item{uuid_column}{uuid column in the dataset. Default is uuid.}

\item{log_name}{character string with name to give to the log}

\item{lower_bound}{lower value of the range (strictly inferior to)}

\item{higher_bound}{higher value of the range (strictly superior to)}
}
\value{
return a list with the dataset checked stored as checked_dataset and
a dataframe with the duration log
}
\description{
Check if value is strictly inferior of the lower threshold or strictly
superior of the higher threshold.
}
\examples{
testdata <- data.frame(
  uuid = c(letters[1:7]),
  duration_audit_start_end_ms = c(
    2475353, 375491, 2654267, 311585, 817270,
    2789505, 8642007
  ),
  duration_audit_start_end_minutes = c(41, 6, 44, 5, 14, 46, 144)
)

check_duration(testdata,
  column_to_check = "duration_audit_start_end_minutes"
)

check_duration(
  testdata,
  column_to_check = "duration_audit_start_end_ms",
  lower_bound = 375490,
  higher_bound = 8642000
)

testdata \%>\%
  check_duration(column_to_check = "duration_audit_start_end_minutes") \%>\%
  check_duration(
    column_to_check = "duration_audit_start_end_ms",
    log_name = "duration_in_ms",
    lower_bound = 375490,
    higher_bound = 8642000
  )
}
