% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_fcs.R
\name{check_fcs}
\alias{check_fcs}
\title{FCS component checks}
\usage{
check_fcs(
  dataset,
  uuid_column = "uuid",
  cereals_column = "fs_fcs_cerealgrainroottuber",
  pulses_column = "fs_fcs_beansnuts",
  dairy_column = "fs_fcs_dairy",
  meat_column = "fs_fcs_meatfishegg",
  vegetables_column = "fs_fcs_vegetableleave",
  fruits_column = "fs_fcs_fruit",
  oil_column = "fs_fcs_fat",
  sugar_column = "fs_fcs_sugar"
)
}
\arguments{
\item{dataset}{dataset to be check as a dataframe or a list with the dataframe stored as "checked_dataset"}

\item{uuid_column}{uuid column in the dataset. Default is uuid.}

\item{cereals_column}{Variable name for- In the last 7 days, on how many days did your household eat cereals, grains, roots and tubers, including wild roots?}

\item{pulses_column}{Variable name for-  In the last 7 days, on how many days did your household eat any beans / legumes, pulses or nuts?}

\item{dairy_column}{Variable name for- In the last 7 days, on how many days did your household drink milk or eat other dairy products?}

\item{meat_column}{Variable name for-  In the last 7 days, on how many days did your household eat meat, fish or eggs?}

\item{vegetables_column}{Variable name for- In the last 7 days, on how many days did your household eat vegetables or leaves, including all wild vegetables and leaves?}

\item{fruits_column}{Variable name for- In the last 7 days, on how many days did your household eat fruits, including all wild fruits?}

\item{oil_column}{Variable name for- In the last 7 days, on how many days did your household eat oil, fat, or butter?}

\item{sugar_column}{Variable name for- In the last 7 days, on how many days did your household eat sugar or sugary foods?}
}
\value{
dataset with potential issues
}
\description{
FCS component checks
}
