% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_others.R
\name{check_others}
\alias{check_others}
\title{Generate a log for other follow up questions}
\usage{
check_others(dataset, uuid_column = "uuid", columns_to_check = NULL)
}
\arguments{
\item{dataset}{dataset to be check as a dataframe or a list with the dataframe stored as "checked_dataset"}

\item{uuid_column}{uuid column in the dataset. Default is "uuid".}

\item{columns_to_check}{Columns to include in the checks}
}
\value{
a log with all the values that may require recoding
}
\description{
Generate a log for other follow up questions
}
\examples{
check_others(
  dataset = cleaningtools::cleaningtools_clean_data,
  uuid_column = "X_uuid",
  columns_to_check = names(cleaningtools::cleaningtools_clean_data |>
    dplyr::select(ends_with("_other")) |>
    dplyr::select(-contains(".")))
)

}
