% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_others.R
\name{check_others_checks}
\alias{check_others_checks}
\title{Check if the input passed to the check_others function is correct}
\usage{
check_others_checks(dataset, uuid_column, columns_to_check)
}
\arguments{
\item{dataset}{dataset to be check as a dataframe or a list with the dataframe stored as "checked_dataset"}

\item{uuid_column}{uuid column in the dataset. Default is uuid.}

\item{columns_to_check}{Columns to include in the checks}
}
\value{
nothing
}
\description{
Check if the input passed to the check_others function is correct
}
