% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_outliers.R
\name{check_outliers}
\alias{check_outliers}
\title{check outliers over the dataset}
\usage{
check_outliers(
  df,
  uuid_col_name = "uuid",
  element_name = "checked_dataset",
  kobo_survey = NULL,
  kobo_choices = NULL,
  cols_to_add_cleaning_log = NULL,
  strongness_factor = 3,
  minimum_unique_value_of_variable = NULL,
  remove_choice_multiple = TRUE,
  sm_sep = ".",
  columns_to_remove = NULL
)
}
\arguments{
\item{df}{data frame or a list}

\item{uuid_col_name}{UUID. Default is uuid}

\item{element_name}{name of the dataset in list}

\item{kobo_survey}{Kobo survey sheet. Default is NULL.}

\item{kobo_choices}{Kobo choices sheet. Default is NULL}

\item{cols_to_add_cleaning_log}{Variables those must be included in the output}

\item{strongness_factor}{Strongness factor define how strong your outliers will be. The default is 3.}

\item{minimum_unique_value_of_variable}{Default is NULL, mean this parameter won't be considered. For example 10 means for any variable where number of unique value is less than 10, then the variable won't be considered for outlier checking.}

\item{remove_choice_multiple}{TRUE (default) will remove choice multiple questions from the output.}

\item{sm_sep}{Separator for choice multiple questions. The default is "."}

\item{columns_to_remove}{Variables those must not consider for outlier checks even though they are numeric}
}
\value{
Outliers
}
\description{
check outliers over the dataset
}
\examples{
df_outlier<- data.frame(
 uuid = paste0("uuid_", 1:100),
 one_value = c(round(runif(90, min = 45,max =55)), round(runif(5)), round(runif(5,99,100))),
 expense = c(sample(200:500,replace = TRUE,size = 95),c(600,100,80,1020,1050)),
 income = c(c(60,0,80,1020,1050),sample(20000:50000,replace = TRUE,size = 95)),
 yy = c(rep(100,99),10)
)

check_outliers(df = df_outlier,uuid_col_name = "uuid")
}
