% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_combined_log.R
\name{create_combined_log}
\alias{create_combined_log}
\title{Merging the cleaning logs}
\usage{
create_combined_log(list_of_log, dataset_name = "checked_dataset")
}
\arguments{
\item{list_of_log}{A list file containig all the logs}

\item{dataset_name}{The dataset's name, if it exists in the list_of_log}
}
\value{
Combined cleaning log
}
\description{
Merging the cleaning logs
}
\examples{
list <- cleaningtools::cleaningtools_raw_data |>
  check_pii(uuid_column = "X_uuid") |>
  check_duplicate(uuid_column = "X_uuid") |>
  check_value(uuid_column = "X_uuid")

create_combined_log(list_of_log = list)

}
