% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_logic_for_other.R
\name{create_logic_for_other}
\alias{create_logic_for_other}
\title{Check for value(s) in the dataset}
\usage{
create_logic_for_other(
  kobo_survey,
  sm_sep = ".",
  compare_with_dataset = FALSE,
  data = NULL
)
}
\arguments{
\item{kobo_survey}{Kobo survey sheet}

\item{sm_sep}{Separator for choice multiple}

\item{compare_with_dataset}{Default FALSE will not compare the applicability of logics with the dataset whereas TRUE will check the applicability of logics.}

\item{data}{data}
}
\value{
A dataframe will all logic.
}
\description{
Check for value(s) in the dataset
}
\examples{
create_logic_for_other(kobo_survey = cleaningtools::cleaningtools_survey,
                       sm_sep = ".",
                       data = cleaningtools::cleaningtools_clean_data,
                       compare_with_dataset = TRUE)

}
