% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/review_others.R
\name{review_others}
\alias{review_others}
\title{Review discrepancy between kobo relevancies and the dataset.}
\usage{
review_others(
  data,
  uuid = "uuid",
  kobo_survey,
  sm_sep = ".",
  variables_to_add = NULL,
  questions_not_to_check = NULL
)
}
\arguments{
\item{data}{dataset}

\item{uuid}{uuid column}

\item{kobo_survey}{Kobo survey sheet}

\item{sm_sep}{Separator for choice multiple}

\item{variables_to_add}{string vector optional, if to add some information to the log (today,village etc.)}

\item{questions_not_to_check}{Questions name which you do not want to check (telephone_number etc.. )}
}
\value{
Any discrepancy between kobo relevancies and the dataset.
}
\description{
Review discrepancy between kobo relevancies and the dataset.
}
\examples{
review_others(data = cleaningtools::cleaningtools_clean_data,
uuid = "X_uuid",kobo_survey = cleaningtools_survey)

}
