% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/review_sample_frame_with_data.R
\name{review_sample_frame_with_dataset}
\alias{review_sample_frame_with_dataset}
\title{Compares the sample frame with the clean data}
\usage{
review_sample_frame_with_dataset(
  sample_frame = sample_frame,
  sample_frame_strata_col,
  sample_frame_target_survey_col = "Total.no.of.HH",
  clean_data = clean_data,
  clean_data_strata_column,
  consent_column = "consent",
  value_for_consent_yes = "yes"
)
}
\arguments{
\item{sample_frame}{A sample frame. It mush have a column for strata and another column defining the number of needed survey per strata.}

\item{sample_frame_strata_col}{Column name for strata in the sample frame}

\item{sample_frame_target_survey_col}{Column name defining number of survey per strata in sample frame.}

\item{clean_data}{Clean data}

\item{clean_data_strata_column}{Column name for strata in clean data}

\item{consent_column}{Dataset consent column}

\item{value_for_consent_yes}{Choice for consent yes}
}
\value{
a dataset with the two additional columns with the number of completed surveys and number of remaining survey.
}
\description{
Compares the sample frame with the clean data
}
\examples{
review_sample_frame_with_dataset(sample_frame = cleaningtools::cleaningtools_sample_frame ,
sample_frame_strata_col = "Neighbourhood",
sample_frame_target_survey_col ="Total.no.of.HH",
clean_data = cleaningtools::cleaningtools_clean_data,
clean_data_strata_column = "neighbourhood",
consent_column = "consent_remote",
value_for_consent_yes = "yes")
}
